actuator_id now works. Removed unnecessary GCS handlers in feed_forward

+++ sw/airborne/modules/fs_landing/actuator_id.c
@@ -9,5 +9,5 @@ float max_delta_amplitude = 0.f;
 
 void add_chirp(struct fs_landing_t *actuator_values) {
     float current_yaw = stateGetNedToBodyEulers_f()->psi;
-    actuator_values->commands[selected_actuator] += max_delta_amplitude * sinf(current_yaw);
+    actuator_values->commands[selected_actuator] += (int32_t) (max_delta_amplitude * sinf(current_yaw));
 }
